move binary into seprate dir, expose public api for helper functions while keeping ui module private
