--- src/main.rs
@@ -1,5 +1,0 @@
-use json_parser::run;
-
-fn main() {
-    run();
-}
